Bring the whole repo to a clean make lint (ruff check .)
88 pre-existing/newly-introduced lint findings -> 0, across src/, scripts/,
and the new training/ tree. Nothing here changes behavior; verified with a
full pytest + mypy --strict pass after every batch, plus a second live
functional re-run (a real sweep, cost profiling, metrics backfill, and
leaderboard generation against synthetic data) after the changes that
touched actual logic rather than just formatting, since those aren't
covered by src/'s test suite.

- pyproject.toml: exclude archive/ (a frozen pre-flatten reference
  snapshot, not maintained code -- linting it further defeats its purpose
  as a faithful historical copy) and add training to the first-party
  import-sort roots, matching src/scripts/tests.
- The 18 D205 (docstring summary/description) findings in src/ predate
  this session's src/ flatten; fixed by collapsing each wrapped summary
  onto one line, shortening a few that didn't fit in 100 columns.
- One real judgment call: left src/adaptivevision/config.py's
  threshold_percentile if/else as-is with a documented noqa rather than
  collapsing it to dict.get() as suggested -- the surrounding comment
  explains a deliberate absent-vs-explicit-null distinction, and silently
  overriding an explained design decision on gated production code isn't
  a call a linter gets to make unilaterally.
- Everything else is line-wrapping, import sorting, an unused-variable
  rename, and one contextlib.suppress simplification -- no behavior change.

+++ training/dashboard_app.py
@@ -290,7 +290,7 @@ def _compute_heatmap(
         ``(None, None, [])`` if the patch grid shape can't be recovered.
     """
     n_patches = patch_features.shape[0]
-    grid = int(round(n_patches**0.5))
+    grid = round(n_patches**0.5)
     if grid * grid != n_patches:
         return None, None, []
